Start in SwiftUI
I started the Xcode project as "Game" rather than "App." However, we
want the latter, as the app will be in SwiftUI, with tests, and the
game will be shown via SpriteView. So, add SwiftUI so that the project
is as if we started with "App," including tests. In a future commit,
add SpriteKit via SpriteView.

Also, I'm trying to get experience with TDD. So, add first test. (Does
"Hello, world!" exist?)