Configure SwiftLint opening brace spacing
Adopting Airbnb's guide for wrapping long function declarations
(https://github.com/airbnb/swift#long-function-declaration). It's
simpler than Google's and creates more readable code.

SwiftLint was warning that "Opening braces should be preceded by a
single space and on the same line as the declaration." We don't want
that for multiline functions. So, configure that rule to allow
multiline functions.